ml-service/requirements.txt
fastapi>=0.115.0,<0.116
uvicorn[standard]>=0.30.0
pydantic-settings>=2.5.0
lightgbm>=4.3.0
scikit-learn>=1.4.0
pandas>=2.1.0
numpy>=1.26.0
category-encoders>=2.6.0
joblib>=1.3.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
